requirements.txt
asgiref==3.5.0
CacheControl==0.12.11
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.5
django-cors-headers==3.11.0
django-environ==0.4.5
django-extensions==3.1.5
django-organizations==2.0.1
djangorestframework==3.13.1
drf-firebase-auth==1.0.0
firebase-admin==4.5.3
google-api-core==1.31.6
google-api-python-client==2.50.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.4
packaging==21.3
Pillow==9.0.1
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==3.0.9
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
